Revert "refactor: expression.tsをlibディレクトリに移動"
This reverts commit 3ccc14abefca42c8c17ee750610e39571b55be25.
